+++ spatial_app/src/components/Navbar.tsx
@@ -2,31 +2,29 @@ import Link from 'next/link';
 
 export default function Navbar(){
     return (
-        <nav className = "bg-gradient-to-l from-[#9B2929] to-[#383E78]">
-          <ul>
+        <nav className="bg-gradient-to-l from-[#9B2929] to-[#383E78]">
+          <ul className="flex justify-end pr-6 py-3">
             <li>
               <Link href="/">
-              <div className="link">Home</div>
+                <div className="link">Home</div>
               </Link>
             </li>
-            <li>
+            <li className="ml-4">
               <Link href="/about">
-              <div className="link">About</div>
+                <div className="link">About</div>
               </Link>
             </li>
-            <li>
+            <li className="ml-4">
               <Link href="/convert_now">
-              <div className="link">Convert Now</div>
+                <div className="link">Convert Now</div>
               </Link>
             </li>
-            <li>
+            <li className="ml-4">
               <Link href="/help">
-              <div className="link">Help</div>
+                <div className="link">Help</div>
               </Link>
             </li>
           </ul>
         </nav>
       );
 }
-
-
