+++ spatial_app/src/pages/index.tsx
@@ -3,14 +3,7 @@ import Link from "next/link";
 import Navbar from '../components/Navbar';
 import { Button } from "../components/ui/button"
 
-const handleClick = () => {
-  console.log("Button Clicked")
-};
-
 export default function Home() {
-  // Debugging statement to ensure the component renders
-  console.log("Home component rendered");
-
   const handleClick = () => {
     console.log("Button Clicked");
   };
@@ -22,9 +15,15 @@ export default function Home() {
         <meta name="description" content="Convert any 2D Video to a Spatial Video"/>
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <main className="flex min-h-screen flex-col items-center justify-center bg-gradient-to-l from-[#9B2929] to-[#383E78]">
-        {/* Debugging statement to ensure the Button component renders */}
-        <h1>Hi</h1>
+      <main className="flex flex-col items-center justify-center bg-gradient-to-l from-[#9B2929] to-[#383E78]">
+        <section className="h-screen flex flex-col items-center justify-center">
+          <h1>Convert your cherished 2D videos to spatial videos for your Apple Vision Pro</h1>
+          {/* You can add more content here */}
+        </section>
+        <section className="h-screen flex flex-col items-center justify-center">
+          <h1>Example Text</h1>
+          {/* You can add more content here */}
+        </section>
       </main>
     </>
   );
